refactor(toolkit): move code interpreter provisioning into builder package
  Move-only commit (byte-identical function bodies) so git blame -C -C -C
  traces every line to its original commit. Reshape follows separately.
